Change PatchSource to retrieve latest patches file
The source for get_patch_list is a raw json file from the CommunityDragon/Data repo. Currently the file being sourced is fixed to a specific commit, so it does not get updated with new patches (and hasn't been since 8.9). By changing the source to the file on master, it will automatically retrieve the latest patch info as long as the CommunityDragon/Data repo stays updated.