Fix item loading for older versions of the game.
In older versions of the League of Legends, e.g. 6.5.1, jungle item
enchantments "colloq" field was set to nill. This went unnoticed through
parser and caused an error when trying to load the items.

Quick repro of a bug:
```python
import cassiopeia as cass

items = Items(region='NA', version='6.5.1')
```

This patch treats nill values in "colloq" field (None in Python) by
ignoring them as if the field was not set at all.